+++ packages/bluesky/lib/src/unspecced/unspecced_service.dart
@@ -10,6 +10,7 @@ import 'package:atproto_core/atproto_core.dart' as core;
 import '../bluesky_base_service.dart';
 import '../entities/feed.dart';
 import '../entities/feed_generators.dart';
+import '../entities/skeleton_actors_by_query.dart';
 import '../entities/skeleton_feed.dart';
 import '../entities/skeleton_posts_by_query.dart';
 
@@ -446,6 +447,126 @@ sealed class UnspeccedService {
     int? limit,
     String? cursor,
   });
+
+  /// Backend Actors (profile) search, returning only skeleton.
+  ///
+  /// ## Parameters
+  ///
+  /// - [query]: search query string; syntax, phrase, boolean, and faceting is
+  ///            unspecified, but Lucene query syntax is recommended
+  ///
+  /// - [typeahead]: If true, acts as fast/simple `typeahead` query.
+  ///
+  /// - [limit]: Maximum number of search results. From 1 to 100.
+  ///            The default is 25.
+  ///
+  /// - [cursor]: Optional pagination mechanism; may not necessarily allow
+  ///             scrolling through entire result set
+  ///
+  /// ## Lexicon
+  ///
+  /// - app.bsky.feed.searchActorsSkeleton
+  ///
+  /// ## Reference
+  ///
+  /// - https://github.com/bluesky-social/atproto/blob/main/lexicons/app/bsky/unspecced/searchActorsSkeleton.json
+  Future<core.XRPCResponse<SkeletonActorsByQuery>> searchActorsByQuerySkeleton(
+    final String query, {
+    bool? typeahead,
+    int? limit,
+    String? cursor,
+  });
+
+  /// Backend Actors (profile) search, returning only skeleton
+  /// as JSON representation.
+  ///
+  /// ## Parameters
+  ///
+  /// - [query]: search query string; syntax, phrase, boolean, and faceting is
+  ///            unspecified, but Lucene query syntax is recommended
+  ///
+  /// - [typeahead]: If true, acts as fast/simple `typeahead` query.
+  ///
+  /// - [limit]: Maximum number of search results. From 1 to 100.
+  ///            The default is 25.
+  ///
+  /// - [cursor]: Optional pagination mechanism; may not necessarily allow
+  ///             scrolling through entire result set
+  ///
+  /// ## Lexicon
+  ///
+  /// - app.bsky.feed.searchActorsSkeleton
+  ///
+  /// ## Reference
+  ///
+  /// - https://github.com/bluesky-social/atproto/blob/main/lexicons/app/bsky/unspecced/searchActorsSkeleton.json
+  Future<core.XRPCResponse<Map<String, dynamic>>>
+      searchActorsByQuerySkeletonAsJson(
+    final String query, {
+    bool? typeahead,
+    int? limit,
+    String? cursor,
+  });
+
+  /// Returns a pagination for backend Actors (profile) search,
+  /// returning only skeleton.
+  ///
+  /// ## Parameters
+  ///
+  /// - [query]: search query string; syntax, phrase, boolean, and faceting is
+  ///            unspecified, but Lucene query syntax is recommended
+  ///
+  /// - [typeahead]: If true, acts as fast/simple `typeahead` query.
+  ///
+  /// - [limit]: Maximum number of search results. From 1 to 100.
+  ///            The default is 25.
+  ///
+  /// - [cursor]: Optional pagination mechanism; may not necessarily allow
+  ///             scrolling through entire result set
+  ///
+  /// ## Lexicon
+  ///
+  /// - app.bsky.feed.searchActorsSkeleton
+  ///
+  /// ## Reference
+  ///
+  /// - https://github.com/bluesky-social/atproto/blob/main/lexicons/app/bsky/unspecced/searchActorsSkeleton.json
+  core.Pagination<SkeletonActorsByQuery> paginateActorsByQuerySkeleton(
+    final String query, {
+    bool? typeahead,
+    int? limit,
+    String? cursor,
+  });
+
+  /// Returns a pagination for backend Actors (profile) search,
+  /// returning only skeleton as JSON representation.
+  ///
+  /// ## Parameters
+  ///
+  /// - [query]: search query string; syntax, phrase, boolean, and faceting is
+  ///            unspecified, but Lucene query syntax is recommended
+  ///
+  /// - [typeahead]: If true, acts as fast/simple `typeahead` query.
+  ///
+  /// - [limit]: Maximum number of search results. From 1 to 100.
+  ///            The default is 25.
+  ///
+  /// - [cursor]: Optional pagination mechanism; may not necessarily allow
+  ///             scrolling through entire result set
+  ///
+  /// ## Lexicon
+  ///
+  /// - app.bsky.feed.searchActorsSkeleton
+  ///
+  /// ## Reference
+  ///
+  /// - https://github.com/bluesky-social/atproto/blob/main/lexicons/app/bsky/unspecced/searchActorsSkeleton.json
+  core.Pagination<Map<String, dynamic>> paginateActorsByQuerySkeletonAsJson(
+    final String query, {
+    bool? typeahead,
+    int? limit,
+    String? cursor,
+  });
 }
 
 final class _UnspeccedService extends BlueskyBaseService
@@ -658,6 +779,65 @@ final class _UnspeccedService extends BlueskyBaseService
         cursor: cursor,
       );
 
+  @override
+  Future<core.XRPCResponse<SkeletonActorsByQuery>> searchActorsByQuerySkeleton(
+    final String query, {
+    bool? typeahead,
+    int? limit,
+    String? cursor,
+  }) async =>
+      await _searchActorsByQuerySkeleton(
+        query: query,
+        typeahead: typeahead,
+        limit: limit,
+        cursor: cursor,
+        to: SkeletonActorsByQuery.fromJson,
+      );
+
+  @override
+  Future<core.XRPCResponse<Map<String, dynamic>>>
+      searchActorsByQuerySkeletonAsJson(
+    final String query, {
+    bool? typeahead,
+    int? limit,
+    String? cursor,
+  }) async =>
+          await _searchActorsByQuerySkeleton(
+            query: query,
+            typeahead: typeahead,
+            limit: limit,
+            cursor: cursor,
+          );
+
+  @override
+  core.Pagination<SkeletonActorsByQuery> paginateActorsByQuerySkeleton(
+    final String query, {
+    bool? typeahead,
+    int? limit,
+    String? cursor,
+  }) =>
+      _paginateActorsByQuerySkeleton(
+        query: query,
+        typeahead: typeahead,
+        limit: limit,
+        cursor: cursor,
+        to: SkeletonActorsByQuery.fromJson,
+      );
+
+  @override
+  core.Pagination<Map<String, dynamic>> paginateActorsByQuerySkeletonAsJson(
+    final String query, {
+    bool? typeahead,
+    int? limit,
+    String? cursor,
+  }) =>
+      _paginateActorsByQuerySkeleton(
+        query: query,
+        typeahead: typeahead,
+        limit: limit,
+        cursor: cursor,
+      );
+
   Future<core.XRPCResponse<T>> _findPopularFeed<T>({
     required bool? includeNsfw,
     required int? limit,
@@ -757,7 +937,7 @@ final class _UnspeccedService extends BlueskyBaseService
     core.To<T>? to,
   }) async =>
       await super.get(
-        'searchSkeletonPosts',
+        'searchPostsSkeleton',
         parameters: _buildSearchPostsSkeletonParams(
           query: query,
           limit: limit,
@@ -773,9 +953,45 @@ final class _UnspeccedService extends BlueskyBaseService
     core.To<T>? to,
   }) =>
       super.paginate(
-        'searchPosts',
+        'searchPostsSkeleton',
         parameters: _buildSearchPostsSkeletonParams(
           query: query,
+          limit: limit,
+          cursor: cursor,
+        ),
+        to: to,
+      );
+
+  Future<core.XRPCResponse<T>> _searchActorsByQuerySkeleton<T>({
+    required String query,
+    required bool? typeahead,
+    required int? limit,
+    required String? cursor,
+    core.To<T>? to,
+  }) async =>
+      await super.get(
+        'searchActorsSkeleton',
+        parameters: _buildSearchActorsSkeletonParams(
+          query: query,
+          typeahead: typeahead,
+          limit: limit,
+          cursor: cursor,
+        ),
+        to: to,
+      );
+
+  core.Pagination<T> _paginateActorsByQuerySkeleton<T>({
+    required String query,
+    required bool? typeahead,
+    required int? limit,
+    required String? cursor,
+    core.To<T>? to,
+  }) =>
+      super.paginate(
+        'searchActorsSkeleton',
+        parameters: _buildSearchActorsSkeletonParams(
+          query: query,
+          typeahead: typeahead,
           limit: limit,
           cursor: cursor,
         ),
@@ -820,6 +1036,19 @@ final class _UnspeccedService extends BlueskyBaseService
   }) =>
       {
         'q': query,
+        'limit': limit,
+        'cursor': cursor,
+      };
+
+  Map<String, dynamic> _buildSearchActorsSkeletonParams({
+    required String query,
+    required bool? typeahead,
+    required int? limit,
+    required String? cursor,
+  }) =>
+      {
+        'q': query,
+        'typeahead': typeahead,
         'limit': limit,
         'cursor': cursor,
       };